fix: retarget Proof path rename leftovers for CI
Regenerate the lockfile for packages/proof and update import, skill, and
tsconfig paths so the directory rename installs and builds on its own.
