+++ packages/explorer/src/web/app/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { ThemeProvider } from './hooks/useTheme';
-import { EffortGraphApp } from './components/EffortGraphApp';
+import { EffortGraphApp } from './components/ProofApp';
 import './globals.css';
 
 const root = document.getElementById('root');
