+++ scripts/bumpVersions.ts
@@ -357,10 +357,10 @@ async function main(): Promise<void> {
     await fs.readFile('packages/flatbread/package.json', 'utf8')
   );
   const effortGraphManifest = JSON.parse(
-    await fs.readFile('packages/effort-graph/package.json', 'utf8')
+    await fs.readFile('packages/proof/package.json', 'utf8')
   );
   await fs.writeFile(
-    'packages/effort-graph/skills/effort-graph/release.json',
+    'packages/proof/skills/proof/release.json',
     `${JSON.stringify(
       {
         format: 1,
